started server-side validation

+++ Controllers/ClubController.cs
@@ -19,6 +19,10 @@ namespace FutManager.Controllers
         }
         public RedirectToActionResult Add(string name,string league,int rating)
         {
+            if(name == null || league == null || rating<1)
+            {
+                return RedirectToAction(actionName: "Index", controllerName:"Error");
+            }
             DataService.AddClub(name, league, rating);
             return RedirectToAction(actionName: "Index");
         }
@@ -28,6 +32,10 @@ namespace FutManager.Controllers
         }
         public RedirectToActionResult Remove(int id,string password)
         {
+            if(!DataService.GetClubs().Any(c => c.Id == id) || id==0)
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             if(password == "password")
             {
                 DataService.DeleteClub(id);
@@ -36,10 +44,18 @@ namespace FutManager.Controllers
         }
         public IActionResult Edit(int id)
         {
+            if(!DataService.GetClubs().Any(y => y.Id == id) || id==0)
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             return View(DataService.GetClubs().FirstOrDefault(x => x.Id == id));
         }
         public RedirectToActionResult EditConfirmed(string name, string league, int rating, string password,int id)
         {
+            if(name==null|| league==null|| rating < 1|| !DataService.GetClubs().Any(y => y.Id == id) || id == 0)
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             if (password == "password")
             {
                 DataService.EditClub(id, name, league, rating);
@@ -48,7 +64,10 @@ namespace FutManager.Controllers
         }
         public IActionResult Details(int id)
         {
-
+            if (!DataService.GetClubs().Any(y => y.Id == id))
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             return View(DataService.GetClubs().FirstOrDefault(x => x.Id == id));
         }
     }
